uiui

+++ Gravy_1.0/EditorViewController.h
@@ -40,7 +40,7 @@ typedef NS_ENUM(NSInteger, KnobId){
     int histLowestValue;
     int histHighestValue;
     int saturationSpline[256];
-    int vibranceSpline[361];
+    float vibranceSpline[361];
     UIImageView* bgImageView;
     UIScrollView* scrollView;
     UIPageControl* pageControl;
